requirements.txt
torch>=1.10.0
torchvision>=0.11.0
streamlit>=1.10.0
Pillow
numpy
